test: replace `u8` with `i32` as the default scalar type

+++ src/dense/convert.rs
@@ -216,23 +216,23 @@ mod tests {
 
     #[test]
     fn test_from() {
-        let expected = Matrix::<u8, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
+        let expected = Matrix::<i32, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
 
         dispatch_unary! {{
-            let seq: [[u8; 3]; 2] = [[1, 2, 3], [4, 5, 6]];
-            let output = Matrix::<u8, O>::from(seq);
+            let seq: [[i32; 3]; 2] = [[1, 2, 3], [4, 5, 6]];
+            let output = Matrix::<i32, O>::from(seq);
             assert_eq!(output, expected);
 
-            let seq: Box<[[u8; 3]; 2]> = Box::new([[1, 2, 3], [4, 5, 6]]);
-            let output = Matrix::<u8, O>::from(seq);
+            let seq: Box<[[i32; 3]; 2]> = Box::new([[1, 2, 3], [4, 5, 6]]);
+            let output = Matrix::<i32, O>::from(seq);
             assert_eq!(output, expected);
 
-            let seq: Box<[[u8; 3]]> = Box::new([[1, 2, 3], [4, 5, 6]]);
-            let output = Matrix::<u8, O>::from(seq);
+            let seq: Box<[[i32; 3]]> = Box::new([[1, 2, 3], [4, 5, 6]]);
+            let output = Matrix::<i32, O>::from(seq);
             assert_eq!(output, expected);
 
-            let seq: Vec<[u8; 3]> = vec![[1, 2, 3], [4, 5, 6]];
-            let output = Matrix::<u8, O>::from(seq);
+            let seq: Vec<[i32; 3]> = vec![[1, 2, 3], [4, 5, 6]];
+            let output = Matrix::<i32, O>::from(seq);
             assert_eq!(output, expected);
         }}
     }
@@ -241,56 +241,55 @@ mod tests {
     fn test_try_from() {
         const MAX: usize = isize::MAX as usize;
 
-        let expected = Matrix::<u8, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
+        let expected = Matrix::<i32, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
 
         dispatch_unary! {{
-            let seq: [Box<[u8; 3]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: [Box<[i32; 3]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Box<[u8; 3]>; 2]> =
-                Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Box<[i32; 3]>; 2]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Box<[u8; 3]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Box<[i32; 3]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Vec<Box<[u8; 3]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Vec<Box<[i32; 3]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: [Box<[u8]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: [Box<[i32]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Box<[u8]>; 2]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Box<[i32]>; 2]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Box<[u8]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Box<[i32]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Vec<Box<[u8]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Vec<Box<[i32]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: [Vec<u8>; 2] = [vec![1, 2, 3], vec![4, 5, 6]];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: [Vec<i32>; 2] = [vec![1, 2, 3], vec![4, 5, 6]];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Vec<u8>; 2]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Vec<i32>; 2]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Box<[Vec<u8>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Box<[Vec<i32>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
-            let seq: Vec<Vec<u8>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
-            let output = Matrix::<u8, O>::try_from(seq).unwrap();
+            let seq: Vec<Vec<i32>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
+            let output = Matrix::<i32, O>::try_from(seq).unwrap();
             assert_eq!(output, expected);
 
             // Unable to cover.
@@ -399,8 +398,7 @@ mod tests {
             let error = Matrix::<(), O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::SizeOverflow);
 
-            let seq: Box<[Vec<()>; 3]> =
-                Box::new([vec![(); MAX], vec![(); MAX], vec![(); MAX]]);
+            let seq: Box<[Vec<()>; 3]> = Box::new([vec![(); MAX], vec![(); MAX], vec![(); MAX]]);
             let error = Matrix::<(), O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::SizeOverflow);
 
@@ -461,47 +459,47 @@ mod tests {
             // let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
             // assert_eq!(error, Error::CapacityOverflow);
 
-            let seq: [Box<[u8]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5])];
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: [Box<[i32]>; 2] = [Box::new([1, 2, 3]), Box::new([4, 5])];
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Box<[Box<[u8]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5])]);
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Box<[Box<[i32]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5])]);
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Box<[Box<[u8]>; 2]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5])]);
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Box<[Box<[i32]>; 2]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5])]);
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Vec<Box<[u8]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5])];
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Vec<Box<[i32]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5])];
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: [Vec<u8>; 2] = [vec![1, 2, 3], vec![4, 5]];
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: [Vec<i32>; 2] = [vec![1, 2, 3], vec![4, 5]];
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Box<[Vec<u8>; 2]> = Box::new([vec![1, 2, 3], vec![4, 5]]);
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Box<[Vec<i32>; 2]> = Box::new([vec![1, 2, 3], vec![4, 5]]);
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Box<[Vec<u8>]> = Box::new([vec![1, 2, 3], vec![4, 5]]);
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Box<[Vec<i32>]> = Box::new([vec![1, 2, 3], vec![4, 5]]);
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
 
-            let seq: Vec<Vec<u8>> = vec![vec![1, 2, 3], vec![4, 5]];
-            let error = Matrix::<u8, O>::try_from(seq).unwrap_err();
+            let seq: Vec<Vec<i32>> = vec![vec![1, 2, 3], vec![4, 5]];
+            let error = Matrix::<i32, O>::try_from(seq).unwrap_err();
             assert_eq!(error, Error::LengthInconsistent);
         }}
     }
 
     #[test]
     fn test_from_iter() {
-        let expected = Matrix::<u8, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
+        let expected = Matrix::<i32, RowMajor>::from_rows([[1, 2, 3], [4, 5, 6]]);
 
         dispatch_unary! {{
             let iter = [[1, 2, 3], [4, 5, 6]];
-            let output = Matrix::<u8, O>::from_iter(iter);
+            let output = Matrix::<i32, O>::from_iter(iter);
             assert_eq!(output, expected);
         }}
     }
@@ -510,37 +508,37 @@ mod tests {
     #[should_panic]
     fn test_from_iter_fails_row_major() {
         let iter = [vec![1, 2, 3], vec![4, 5]];
-        Matrix::<u8, RowMajor>::from_iter(iter);
+        Matrix::<i32, RowMajor>::from_iter(iter);
     }
 
     #[test]
     #[should_panic]
     fn test_from_iter_fails_col_major() {
         let iter = [vec![1, 2, 3], vec![4, 5]];
-        Matrix::<u8, ColMajor>::from_iter(iter);
+        Matrix::<i32, ColMajor>::from_iter(iter);
     }
 
     #[test]
     fn test_into() {
         dispatch_unary! {{
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Box<[Box<[u8]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output: Box<[Box<[u8]>]> = matrix.into();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Box<[Box<[i32]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output: Box<[Box<[i32]>]> = matrix.into();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Vec<Box<[u8]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output: Vec<Box<[u8]>> = matrix.into();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Vec<Box<[i32]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output: Vec<Box<[i32]>> = matrix.into();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Box<[Vec<u8>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
-            let output: Box<[Vec<u8>]> = matrix.into();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Box<[Vec<i32>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
+            let output: Box<[Vec<i32>]> = matrix.into();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Vec<Vec<u8>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
-            let output: Vec<Vec<u8>> = matrix.into();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Vec<Vec<i32>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
+            let output: Vec<Vec<i32>> = matrix.into();
             assert_eq!(output, expected);
         }}
     }
