+++ src/dense/convert/into_rows.rs
@@ -78,24 +78,24 @@ mod tests {
     #[test]
     fn test_into_rows() {
         dispatch_unary! {{
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Box<[Box<[u8]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
-            let output: Box<[Box<[u8]>]> = matrix.into_rows();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Box<[Box<[i32]>]> = Box::new([Box::new([1, 2, 3]), Box::new([4, 5, 6])]);
+            let output: Box<[Box<[i32]>]> = matrix.into_rows();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Vec<Box<[u8]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
-            let output: Vec<Box<[u8]>> = matrix.into_rows();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Vec<Box<[i32]>> = vec![Box::new([1, 2, 3]), Box::new([4, 5, 6])];
+            let output: Vec<Box<[i32]>> = matrix.into_rows();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Box<[Vec<u8>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
-            let output: Box<[Vec<u8>]> = matrix.into_rows();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Box<[Vec<i32>]> = Box::new([vec![1, 2, 3], vec![4, 5, 6]]);
+            let output: Box<[Vec<i32>]> = matrix.into_rows();
             assert_eq!(output, expected);
 
-            let matrix = Matrix::<u8, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
-            let expected: Vec<Vec<u8>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
-            let output: Vec<Vec<u8>> = matrix.into_rows();
+            let matrix = Matrix::<i32, O>::from_rows([[1, 2, 3], [4, 5, 6]]);
+            let expected: Vec<Vec<i32>> = vec![vec![1, 2, 3], vec![4, 5, 6]];
+            let output: Vec<Vec<i32>> = matrix.into_rows();
             assert_eq!(output, expected);
         }}
     }
